change usage of vuex store
