+++ src/main.js
@@ -9,9 +9,9 @@ import '~/assets/css/style.less'
 import '~/assets/css/variables.less'
 // import jquery from 'jquery'
 import Vuex from 'vuex'
-import getters from './store/getters'
-import mutations from './store/mutations'
-import actions from './store/actions'
+import articles from './store/articles'
+import ux from './store/ux'
+import links from './store/links'
 
 export default function (Vue, { head, appOptions }) {
   Vue.use(Bootstrap)
@@ -45,32 +45,10 @@ export default function (Vue, { head, appOptions }) {
 
   // set Vuex store
   appOptions.store = new Vuex.Store({
-    state: {
-      allArticles: [],
-      articles: [],
-      language: 'en',
-      languages: {
-        'en': 'English',
-        'it': 'Italiano',
-        'ar': 'العربية',
-        'fa': 'فارسی',
-        'es': 'Español',
-        'fr': 'français',
-        'cn': '中文',
-        'de': 'Deutsch',
-        'ru': 'русский',
-        'hi': 'हिन्दी'
-      },
-      social_links: [
-        // { id: 1, icon: "twitter" },
-        { id: 2, icon: "facebook", url:"https://www.facebook.com/The-Dawn-Breakers-Challenge-329820570524376/" },
-        // { id: 3, icon: "instagram" },
-        // { id: 4, icon: "pinterest" },
-        { id: 5, icon: "youtube", url:"" }
-      ]
-    },
-    mutations,
-    getters,
-    actions
+    modules: {
+      a: articles,
+      b: ux,
+      c: links
+    }
   })
 }
